moved ThemeConfiguration to same package as Theme and ThemeManager
