+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/controller/ControllerConfiguration.java
@@ -18,7 +18,7 @@
 
 package org.phoenicis.javafx.controller;
 
-import org.phoenicis.javafx.views.ThemeConfiguration;
+import org.phoenicis.javafx.views.common.ThemeConfiguration;
 import org.phoenicis.library.LibraryConfiguration;
 import org.phoenicis.settings.SettingsConfiguration;
 import org.phoenicis.apps.AppsConfiguration;
